Update README.md
Algolia provides 2 types of key:

- search only api key
- admin api key

In this context, we need the latter to create/update records. However it's not meant to be exposed to public (like on frontend code, or in GitHub repo). So we recommend people to put it in the environment variable and let them access it like this.

What do you think?